requirements.txt
asgiref==3.7.2
comtypes==1.2.0
Django==4.2.3
django-cors-headers==4.1.0
djangorestframework==3.14.0
dnspython==2.3.0
joblib==1.3.1
numpy==1.25.0
pandas==2.0.3
pymongo==4.4.0
pypiwin32==223
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.1.0
typing_extensions==4.7.1
tzdata==2023.3